started the FeedPageActivity
