+++ app/src/main/java/info/adrian/powerplant/AddPostActivity.java
@@ -157,23 +157,4 @@ public class AddPostActivity extends AppCompatActivity {
             }
         });
     }
-
-    //display the list of posts
-    private void queryPosts() {
-        ParseQuery<Post> query = ParseQuery.getQuery(Post.class);
-
-        query.include(Post.KEY_USER);
-        query.findInBackground(new FindCallback<Post>() {
-            @Override
-            public void done(List<Post> posts, ParseException e) {
-                if(e != null){
-                    Log.e(TAG, "Issue with getting posts", e);
-                    return;
-                }
-                for(Post post : posts){
-                    Log.i(TAG, "Post: " + post.getDescription() + ", username:" + post.getUser().getUsername());
-                }
-            }
-        });
-    }
 }
