+++ app/src/main/java/info/adrian/powerplant/FeedPageActivity.java
@@ -1,17 +1,28 @@
 package info.adrian.powerplant;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.google.android.material.bottomappbar.BottomAppBar;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.parse.FindCallback;
+import com.parse.ParseException;
+import com.parse.ParseQuery;
+
+import java.util.List;
 
 public class FeedPageActivity extends AppCompatActivity {
 
@@ -20,6 +31,9 @@ public class FeedPageActivity extends AppCompatActivity {
     private TextView test;
     private BottomAppBar mBottomAppBar;
     private FloatingActionButton mAddPost;
+    private RecyclerView posts;
+
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -36,6 +50,26 @@ public class FeedPageActivity extends AppCompatActivity {
             public void onClick(View v) {
                 Intent postActivity = new Intent(getApplicationContext(), AddPostActivity.class);
                 startActivity(postActivity);
+            }
+        });
+
+        queryPosts();
+    }
+
+    private void queryPosts() {
+        ParseQuery<Post> query = ParseQuery.getQuery(Post.class);
+
+        query.include(Post.KEY_USER);
+        query.findInBackground(new FindCallback<Post>() {
+            @Override
+            public void done(List<Post> posts, ParseException e) {
+                if(e != null){
+                    Log.e(TAG, "Issue with getting posts", e);
+                    return;
+                }
+                for(Post post : posts){
+                    Log.i(TAG, "Post: " + post.getDescription() + ", username:" + post.getUser().getUsername());
+                }
             }
         });
     }
